Move waitForDeviceToReboot helper to device-util
