+++ src/cmd/device-util.js
@@ -1,3 +1,4 @@
+const usbUtils = require('./usb-util');
 /**
  * Check if the string can represent a valid device ID.
  *
@@ -45,3 +46,28 @@ module.exports.getDevice = ({ id, api, auth, displayName = null, dontThrow = fal
 			throw error;
 		});
 };
+
+/**
+* Waits for the device to reboot to reboot by checking if the device is ready to accept control requests.
+* It waits for a maximum of 60 seconds with a 1-second interval.
+*/
+module.exports.waitForDeviceToReboot = async (deviceId) => {
+	const REBOOT_TIME_MSEC = 60000;
+	const REBOOT_INTERVAL_MSEC = 1000;
+	const start = Date.now();
+	while (Date.now() - start < REBOOT_TIME_MSEC) {
+		try {
+			await _delay(REBOOT_INTERVAL_MSEC);
+			const device = await usbUtils.reopenDevice({ id: deviceId });
+			// Waiting for any control request to work to ensure the device is ready
+			await device.getDeviceId();
+			return device;
+		} catch (error) {
+			// ignore error
+		}
+	}
+};
+
+async function _delay(ms){
+	return new Promise((resolve) => setTimeout(resolve, ms));
+}
